--- hw04/JoesAutomotive6.4/Form1.cs
@@ -1,10 +1,0 @@
-namespace JoesAutomotive6._4
-{
-    public partial class Form1 : Form
-    {
-        public Form1()
-        {
-            InitializeComponent();
-        }
-    }
-}
