--- hw04/JoesAutomotive6.4/JoesAutomotive6.4.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.7.34031.279
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JoesAutomotive6.4", "JoesAutomotive6.4.csproj", "{5EEB9DBB-D256-481B-9500-97837B523EFA}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5EEB9DBB-D256-481B-9500-97837B523EFA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5EEB9DBB-D256-481B-9500-97837B523EFA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5EEB9DBB-D256-481B-9500-97837B523EFA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5EEB9DBB-D256-481B-9500-97837B523EFA}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {80F94170-6803-49D2-8ED1-69038890CA26}
-	EndGlobalSection
-EndGlobal
