+++ hw04/RockPaperScissors6.9/6.9RockPaperScissors.csproj
@@ -3,7 +3,9 @@
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
     <TargetFramework>net6.0-windows</TargetFramework>
-    <RootNamespace>_6._9RockPaperScissors</RootNamespace>
+    <RootNamespace>
+		namespace RockPaperScissors_6._9
+	</RootNamespace>
     <Nullable>enable</Nullable>
     <UseWindowsForms>true</UseWindowsForms>
     <ImplicitUsings>enable</ImplicitUsings>
