--- hw04/RockPaperScissors6.9/RockPaperScissors6.9.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.7.34031.279
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RockPaperScissors6.9", "RockPaperScissors6.9.csproj", "{19FD9D55-D68A-4838-B3C2-C7A1E59A0918}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{19FD9D55-D68A-4838-B3C2-C7A1E59A0918}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{19FD9D55-D68A-4838-B3C2-C7A1E59A0918}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{19FD9D55-D68A-4838-B3C2-C7A1E59A0918}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{19FD9D55-D68A-4838-B3C2-C7A1E59A0918}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {715BA89D-14D8-47A4-9E8A-6E1356474257}
-	EndGlobalSection
-EndGlobal
